feat: 质量红线支持stage准入准出拦截 #4863 add check times param

+++ src/backend/ci/core/quality/biz-quality/src/main/kotlin/com/tencent/devops/quality/service/v2/QualityHistoryService.kt
@@ -449,7 +449,7 @@ class QualityHistoryService @Autowired constructor(
         interceptList: String,
         createTime: LocalDateTime,
         updateTime: LocalDateTime
-    ): Long {
+    ): Int {
         return historyDao.create(
             dslContext = dslContext,
             projectId = projectId,
